create new package dataset for all legacy/geolite2 handling
