+++ handler/dataset/geolite2-dataset.go
@@ -1,4 +1,4 @@
-package handler
+package dataset
 
 import (
 	"context"
@@ -9,50 +9,9 @@ import (
 	"github.com/m-lab/annotation-service/loader"
 	"github.com/m-lab/annotation-service/parser"
 	"github.com/m-lab/annotation-service/search"
-
-	"google.golang.org/api/iterator"
-
-	"cloud.google.com/go/storage"
 )
 
-// PopulateLatestData will search to the latest Geolite2 files
-// available in GCS and will use them to create a new GeoDataset which
-// it will place into the global scope as the latest version. It will
-// do so safely with use of the currentDataMutex RW mutex. It it
-// encounters an error, it will halt the program.
-func PopulateLatestData() {
-	filename, err := determineFilenameOfLatestGeolite2File()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	CurrentGeoDataset.AddDataset(filename)
-	Geolite2Dataset.Init()
-	LegacyDataset.Init()
-}
-
-// determineFilenameOfLatestGeolite2File will get a list of filenames
-// from GCS and search through them, eventually returing either the
-// latest filename or an error.
-func determineFilenameOfLatestGeolite2File() (string, error) {
-	ctx := context.Background()
-	client, err := storage.NewClient(ctx)
-	if err != nil {
-		return "", err
-	}
-	prospectiveFiles := client.Bucket(BucketName).Objects(ctx, &storage.Query{Prefix: MaxmindPrefix})
-	filename := ""
-	for file, err := prospectiveFiles.Next(); err != iterator.Done; file, err = prospectiveFiles.Next() {
-		if err != nil {
-			return "", err
-		}
-		if file.Name > filename && GeoLite2Regex.MatchString(file.Name) {
-			filename = file.Name
-		}
-
-	}
-	return filename, nil
-}
+const MaxmindPrefix = "Maxmind/" // Folder containing the maxmind files
 
 // LoadGeoLite2Dataset load the Geolite2 dataset with filename from bucket.
 func LoadGeoLite2Dataset(filename string, bucketname string) (*parser.GeoDataset, error) {
