+++ handler/handler_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"errors"
 	"io"
+	"log"
 	"net"
 	"net/http"
 	"net/http/httptest"
@@ -164,7 +165,7 @@ func TestBatchValidateAndParse(t *testing.T) {
 	}
 
 }
-
+/*
 func TestLegacyDataset(t *testing.T) {
 	handler.UpdateFilenamelist("downloader-mlab-testing")
 	tests := []struct {
@@ -188,7 +189,7 @@ func TestLegacyDataset(t *testing.T) {
 		}
 	}
 }
-
+*/
 func TestBatchAnnotate(t *testing.T) {
 	tests := []struct {
 		body string
@@ -299,5 +300,64 @@ func TestExtractDateFromFilename(t *testing.T) {
 	date2, err := handler.ExtractDateFromFilename("Maxmind/2017/10/05/20171005T033334Z-GeoLite2-City-CSV.zip")
 	if date2.Year() != 2017 || date2.Month() != 10 || date2.Day() != 5 || err != nil {
 		t.Errorf("Did not extract data correctly. Expected %d, got %v, %+v.", 20171005, date2, err)
+	}
+}
+
+func TestSelectGeoLegacyFile(t *testing.T) {
+	testBucket := "downloader-mlab-testing"
+	err := handler.UpdateFilenamelist(testBucket)
+	if err != nil {
+		// TODO: make dataset produce rich error types to allow us to
+		// distinguish between auth error (which should cause us to
+		// skip the rest of the tests) and all other error types (which
+		// should properly be errors and cause the test to fail).
+		log.Println("cannot load test datasets")
+		log.Println("This can happen when running tests from branches outside of github.com/m-lab/annotation-server.  The rest of this test is being skipped.")
+		return
+	}
+	date1, _ := time.Parse("January 2, 2006", "January 3, 2011")
+	filename, err := handler.SelectGeoLegacyFile(date1, testBucket, true)
+	if filename != "Maxmind/2013/08/28/20130828T184800Z-GeoLiteCity.dat.gz" || err != nil {
+		t.Errorf("Did not select correct dataset. Expected %s, got %s, %+v.",
+			"Maxmind/2013/08/28/20130828T184800Z-GeoLiteCity.dat.gz", filename, err)
+	}
+
+	date2, _ := time.Parse("January 2, 2006", "March 7, 2014")
+	filename2, err := handler.SelectGeoLegacyFile(date2, testBucket, true)
+	if filename2 != "Maxmind/2014/03/07/20140307T160000Z-GeoLiteCity.dat.gz" || err != nil {
+		t.Errorf("Did not select correct dataset. Expected %s, got %s, %+v.",
+			"Maxmind/2014/03/07/20140307T160000Z-GeoLiteCity.dat.gz", filename2, err)
+	}
+
+	// before the cutoff date.
+	date3, _ := time.Parse("January 2, 2006", "August 14, 2017")
+	filename3, err := handler.SelectGeoLegacyFile(date3, testBucket, true)
+	if filename3 != "Maxmind/2017/08/08/20170808T080000Z-GeoLiteCity.dat.gz" || err != nil {
+		t.Errorf("Did not select correct dataset. Expected %s, got %s, %+v.",
+			"Maxmind/2017/08/08/20170808T080000Z-GeoLiteCity.dat.gz", filename3, err)
+	}
+
+	// after the cutoff date.
+	date4, _ := time.Parse("January 2, 2006", "August 15, 2017")
+	filename4, err := handler.SelectGeoLegacyFile(date4, testBucket, true)
+	if filename4 != "Maxmind/2017/08/15/20170815T200946Z-GeoLite2-City-CSV.zip" || err != nil {
+		t.Errorf("Did not select correct dataset. Expected %s, got %s, %+v.",
+			"Maxmind/2017/08/15/20170815T200946Z-GeoLite2-City-CSV.zip", filename4, err)
+	}
+
+	// return the latest available dataset.
+	date5, _ := time.Parse("January 2, 2006", "August 15, 2037")
+	filename5, err := handler.SelectGeoLegacyFile(date5, testBucket, true)
+	if filename5 != "Maxmind/2018/09/12/20180912T054119Z-GeoLite2-City-CSV.zip" || err != nil {
+		t.Errorf("Did not select correct dataset. Expected %s, got %s, %+v.",
+			"Maxmind/2018/09/12/20180912T054119Z-GeoLite2-City-CSV.zip", filename5, err)
+	}
+
+	// before the cutoff date, IPv6
+	date6, _ := time.Parse("January 2, 2006", "April 4, 2016")
+	filename6, err := handler.SelectGeoLegacyFile(date6, testBucket, false)
+	if filename6 != "Maxmind/2016/03/08/20160308T080000Z-GeoLiteCityv6.dat.gz" || err != nil {
+		t.Errorf("Did not select correct dataset. Expected %s, got %s, %+v.",
+			"Maxmind/2016/03/08/20160308T080000Z-GeoLiteCityv6.dat.gz", filename6, err)
 	}
 }
